feat: compilation exceptions implemented

+++ common/src/main/java/ru/practicum/dto/compilation/NewCompilationDto.java
@@ -5,8 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.practicum.UniqueElements;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.List;
@@ -16,7 +14,7 @@ import java.util.List;
 @NoArgsConstructor
 public class NewCompilationDto {
     @UniqueElements
-    List<Integer> events;
+    List<Long> events;
     Boolean pinned = false;
     @Size(min = 1, max = 50)
     @NotNull
